+++ dev/mcp-server-examples/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @transcend-io/mcp-server-examples
 
+## 0.0.17
+
+### Patch Changes
+
+- 2a6a955: Fixes a lot of Sombra tools
+- Updated dependencies [2a6a955]
+- Updated dependencies [557a80b]
+  - @transcend-io/mcp-server-base@1.8.0
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies [5b97f8e]
+  - @transcend-io/mcp-server-base@1.7.4
+
 ## 0.0.15
 
 ### Patch Changes